# fix(ext/web): Aggregate small packets for Resource implementation of ReadableStream