--- a/ext/web/stream_resource.rs
+++ b/ext/web/stream_resource.rs
@@ -1,5 +1,5 @@
-use bytes::BytesMut;
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+use bytes::BytesMut;
 use deno_core::error::type_error;
 use deno_core::error::AnyError;
 use deno_core::op2;